requirements.txt
#!/bin/bash

pandas
numpy
pymongo
mysql-conector-python